// frontend/src/api/models/LeadStage.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type LeadStage =
    | 'new'
    | 'qualifying'
    | 'qualified'
    | 'quoted'
    | 'tentative_booking'
    | 'booked'
    | 'shot'
    | 'delivered'
    | 'closed_lost';