// frontend/src/api/models/BudgetBand.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export const BudgetBand = {
    _10K: '<10k',
    _10K_20K: '10k-20k',
    _20K_40K: '20k-40k',
    _40K: '>40k',
    UNKNOWN: 'unknown',
} as const;

export type BudgetBand = typeof BudgetBand[keyof typeof BudgetBand];